refactored and fixed idk what...
