+++ app.js
@@ -3,7 +3,7 @@ import bodyParser from 'body-parser';
 import cors from  'cors';
 import path from  'path';
 import logger from  'morgan';
-import config from  './config'; // // get our config file
+import config from './database/config'; // // get our config file
 import Swagger from './services/swagger.service';
 import {swaggerRoute, userRouter, taskRouter,homeRouter,shareTaskRouter} from './routers/index.routing';
 import http from 'http';
