--- config.js
@@ -1,14 +1,0 @@
-/**
- * Created by Krys on 29/01/2018.
- */
-// module.exports = {
-//     'secret': 'ilovescotchyscotch',
-// };
-const config = {
-    //run for the docker container
-    // mongoUrl : 'mongodb://mongo:27017/websecurity'
-    // mongoUrl : 'mongodb://localhost:27017/nodejs_summer_exam'
-    mongoUrl: 'mongodb://krystian:yoman123@ds119160.mlab.com:19160/heroku_lv1rbkzt'
-};
-// const url = 'mongodb://websiteWatcher:websiteWatcher123@ds251827.mlab.com:51827/heroku_7l6c835k';
-export  default config;
