+++ database/config.js
@@ -1,0 +1,5 @@
+const config = {
+    // mongoUrl : 'mongodb://localhost:27017/nodejs_summer_exam'
+    mongoUrl: 'mongodb://krystian:yoman123@ds119160.mlab.com:19160/heroku_lv1rbkzt'
+};
+export  default config;
