+++ models/user.js
@@ -20,7 +20,7 @@ class UserModel{
             },
             email: {
                 type: String,
-                validate: /^[a-z0-9._@]{5,}$/,
+                validate: /^[a-zA-Z0-9._@+]{5,}$/,
                 minLength: 5,
                 maxLength: 40,
                 required: true,
