+++ routers/home.routing.js
@@ -1,11 +1,7 @@
-/**
- * Created by Palko on 03/03/2018.
- */
 import express from 'express';
-let homeRouter = express.Router();
-import jwt from 'jsonwebtoken';
-import userController from '../controllers/user.controller';
 import {checkTokenValidity} from '../services/token.service';
+
+let homeRouter = express.Router();
 
 homeRouter.post('',checkTokenValidity, function(req, res) {
         if (req.user) {
@@ -14,7 +10,6 @@ homeRouter.post('',checkTokenValidity, function(req, res) {
         }
        return res.sendStatus(310);
 });
-
 
 
 export default homeRouter;
