+++ routers/index.routing.js
@@ -4,8 +4,6 @@ import taskRouter from "./task.routing";
 import homeRouter from "./home.routing";
 import shareTaskRouter from "./shareTask.routing";
 
-
-
 export {
     swaggerRoute,
     userRouter,
