+++ routers/user.routing.js
@@ -4,9 +4,11 @@ import {validateCaptcha} from '../services/captcha.service';
 import randomString from 'randomstring';
 import fs from 'fs';
 import {checkTokenValidity} from "../services/token.service";
-import {role} from "../services/roles.service";
 
-//define router
+/**
+ * define the router
+ * @type {Router|router|*}
+ */
 let userRouter = express.Router();
 
 /**
@@ -56,7 +58,7 @@ let userRouter = express.Router();
 
 /**
  * @swagger
- * /users:
+ * /users/:
  *  get:
  *      tags:
  *      - user
@@ -72,12 +74,17 @@ let userRouter = express.Router();
  *              description: ok
  *
  */
-userRouter.get('/', [checkTokenValidity, role('admin')], function(req, res) {
-    userController.getAll().then((users, err)=>{
-        if(err)
-            console.error(err);
-        res.json({users: userController.decryptUsersEmail(users)});
-    });
+userRouter.get('/', checkTokenValidity, function(req, res) {
+    if(req.user.role ==='admin'){
+        userController.getAll().then((users, err)=>{
+            if(err)
+                console.error(err);
+            res.json({users: userController.decryptUsersEmail(users)});
+        });
+    }else{
+        res.status(401);
+        res.send('you are not an admin');
+    }
 });
 /**
  * @swagger
@@ -158,16 +165,15 @@ userRouter.post('/signupDev', (req, res)=> {
  *              description: ok
  *
  */
-
 userRouter.post('/login', (req, res)=>{
     userController.loginUser(req.body).then((data)=>{
         res.cookie("thisisTestCookie", "values", { secure:false, maxAge:120000, httpOnly: true })
             .json({
-            success: true,
-            message: 'Enjoy your token!',
-            token: data.token,
-            user: data.user
-        });
+                success: true,
+                message: 'Enjoy your token!',
+                token: data.token,
+                user: data.user
+            });
 
     }, (err)=>{
         if(err.lock){
@@ -196,7 +202,6 @@ userRouter.post('/login', (req, res)=>{
  *
  */
 
-//Rout that checks if the user exists & send email with password reset
 userRouter.post('/forgot', (req, res)=>{
     userController.forgotPassword(req).then((data)=>{
         res.send(data);
@@ -211,14 +216,11 @@ userRouter.post('/forgot', (req, res)=>{
 
 //Link from the email checks if the user who wanted to change the password exists and the token is not expired
 userRouter.get('/reset/:token', function(req, res) {
-
     userController.findUserByToken(req).then((user)=>{
-
         if (!user) {
             //If user does not exist navigate to forgot password form
             return res.redirect('https://angular.sevenamstudio.com/#/forgot');
-    }
-
+        }
         //else navigate to rest form where user can type new password
         return res.redirect('http://' + 'angular.sevenamstudio.com/#' + '/reset/'+ req.params.token);
     });
@@ -244,7 +246,7 @@ userRouter.post('/image', (req, res) => {
     let base64Data = req.body.src.replace(/^data:image\/png;base64,/, "");
     const file = './images/profiles/'+ image;
     fs.writeFile(file, base64Data, 'base64', function(err) {
-      //  console.log(err, '<----err');
+        //  console.log(err, '<----err');
     });
     res.send({imgUrl: image})
 });
