+++ services/token.service.js
@@ -5,8 +5,7 @@ let client='http://localhost:4200';
 
 export function checkTokenValidity(req,res,next) {
     // check header or url parameters or post parameters for token
-    var token = req.body.token || req.query.token || req.headers['x-access-token'];
-
+    let token = req.body.token || req.query.token || req.headers['x-access-token'];
 
     // decode token
     if (token && checkRefererAndOrigin) {
@@ -14,12 +13,8 @@ export function checkTokenValidity(req,res,next) {
         jwt.verify(token, ('superDuperSecretKey'), function (err, decoded) {
             if (err) {
                 return res.status(400).send("Failed to authenticate token.");
-
-
             } else if (decoded.exp <= Date.now()) {
-
                 return res.status(498).send('Access token has expired');
-
             }
             userController.findUserByEmailInToken(decoded.email).then((user)=>{
                 if(!user){
@@ -39,7 +34,6 @@ export function checkTokenValidity(req,res,next) {
 }
 
  function checkRefererAndOrigin(req){
-
      if(req.headers['referer'] == client+"/" && req.headers['origin'] == client){
          return true;
      }
@@ -48,9 +42,8 @@ export function checkTokenValidity(req,res,next) {
 
 export function generateToken(user) {
     const payload = {
-        //todo add roles
         email: user.email
     };
-    //time can be easly change
+    //time can be changed
     return jwt.sign(payload, 'superDuperSecretKey',  { expiresIn: Date.now()+5000 });
 }
